update version indev 1.2.8
Changelog:

- Fixed video playing on Android mobile (no support for iOS yet)
   * Fixed video trailers immediately skipped
   * Fixed displaying scroll after playing scroll video

- Updated the Cook Book System
   * Fixed displaying supposedly future pages when on previous round
   * Inserted the dish frame inside the cookbook
   * Added map that highlights where the dish is located
   * Updated the interface and its design

- Adjusted the cost of each province
   * prev: 0 -> 40 -> 75 -> 115 -> 150 -> 200
   * now: 0 -> 50 -> 80 -> 115 -> 160 -> 200

- Updated the formula for timeDuration
   * KitchenMode prev:  timeDuration = 8 + (no. of ingredients * 4)
   * KitchenMode now:  timeDuration = 7.5 + (no. of ingredients * 2.5)
   * RestaurantMode prev:  timeDuration = 5 + (no. of ingredients * 2.5)
   * RestaurantMode now:  timeDuration = 4 + (no. of ingredients * 2)

- Changed few variable and method names
- Fixed few minor bugs

+++ Visayan-Sojourn/Assets/Scripts/PlayerCoins.cs
@@ -100,7 +100,7 @@ public class PlayerCoins : MonoBehaviour
         _updateDisplayMain.UpdateDisplayCoinBag();
 
         // Check if the coin generated is less than the maximum
-        if (coinsGenerated < coinGeneratorMax)
+        if ((int)coinsGenerated < coinGeneratorMax)
         {
             if (!inCooldown)
             {
